seperate out openweathermap functionality
